feat: register multiple tasks
